test/pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bedrock-agentcore>=1.1.1",
    "dotenv>=0.9.9",
    "google-genai>=1.56.0",
    "strands-agents-tools>=0.2.18",
    "strands-agents[gemini]>=1.20.0",
]
